backend/requirements.txt
# Django and REST framework
Django==5.2.1
djangorestframework==3.14.0

# CORS handling
django-cors-headers==4.3.1

# Static file serving
whitenoise==6.6.0

# Database
# SQLite is included with Python, but if you need PostgreSQL later:
# psycopg2-binary==2.9.9

# Development tools
python-dotenv==1.0.1

# Production server
gunicorn==21.2.0

# Security
django-environ==0.11.2

# Image processing (if you're handling image uploads)
Pillow==10.2.0

# API documentation
drf-yasg==1.21.7

# Testing
pytest==8.0.0
pytest-django==4.8.0

# Code quality
black==24.1.1
flake8==7.0.0
isort==5.13.2 